fix(zkp): restore Windows scratch export and fix UBSan/ASan tests
Keep SECP256K1_API on scratch definitions inside extern "C" so MSVC
DLLs export them without g++ mangling, shift 1 as size_t before 31
to avoid signed overflow, and use stack buffers in multi-party
bulletproof tests instead of leaked mallocs.

+++ src/modules/schnorrsig_mw/tests_impl.h
@@ -111,7 +111,7 @@ void test_schnorrsig_mw_api(secp256k1_scratch *scratch) {
     CHECK(ecount == 3);
     CHECK(secp256k1_schnorrsig_mw_verify_batch(vrfy, scratch, &sigptr, &msgptr, &pkptr, (size_t)1 << (sizeof(size_t)*8-1)) == 0);
     CHECK(ecount == 4);
-    CHECK(secp256k1_schnorrsig_mw_verify_batch(vrfy, scratch, &sigptr, &msgptr, &pkptr, 1 << 31) == 0);
+    CHECK(secp256k1_schnorrsig_mw_verify_batch(vrfy, scratch, &sigptr, &msgptr, &pkptr, (size_t)1 << 31) == 0);
     CHECK(ecount == 5);
 
     secp256k1_context_destroy(none);
